[MLIR][OpenMP] NFC: Split OpenMP dialect definitions
This patch splits definitions for the OpenMP dialect into multiple files to
simplify the addition of new features, reduce merge conflicts, make it easier
to understand, etc. The split is based on the structure of the more mature
LLVMIR dialect. More specifically:

- The OpenMP dialect definition is located in OpenMPDialect.td.
- Base classes for OpenMP operations and types, as well as generic OpenMP types
are moved to OpenMPOpBase.td.
- OpenMP enumeration attributes, their case attributes and shared base classes
for these are placed in OpenMPEnums.td.
- Other OpenMP attributes are separated into OpenMPAttrDefs.td.
- OpenMPOps.td only contains operation definitions.

Even though this change should be useful on its own, it is intended as a
precursor to a follow-up PR in which operation arguments and attributes are
split into clause-specific classes which are then shared by all operations to
which they apply to. Without this prior change, that approach would make the
OpenMPOps.td file harder to navigate.